Add variables for scheme, host & port
Declaring and using a variable for the forward_host in the proxy_pass directive means that Nginx will no longer fail to start if that host is down. A 503 will be returned for any requests instead.
I have also added variables for the forward_scheme and forward_port elements also as these could be useful variables to have access to for any custom elements users may wish to add.
I have not declared a variable for forward_path as it is not mandatory and an empty value would cause the set directive to fail.